Move GNN utility functions, Scheduler and Trainer class out of main and into separate files
